requirements.txt
cassandra_driver==3.14.0
libMHCUDA==2.1.0
python-igraph==0.7.1.post6
jinja2==2.10
sourced-ml[tf]==0.6.1
